Rename module.
